requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pillow==10.2.0
discord.py==2.3.2